サイドバー(cycleのabout)を別ファイルに切り出し
